pyproject.toml
[tool.poetry]
name = "sp-crawler-job"
version = "0.1.0"
description = ""
authors = ["elviawu <elviawu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
requests = "^2.32.3"
pydantic = "^2.9.2"
loguru = "^0.7.2"
beautifulsoup4 = "^4.12.3"
selenium = "^4.26.0"
webdriver-manager = "^4.0.2"
fake-useragent = "^1.5.1"
lxml = "^5.3.0"
aiohttp = "^3.10.10"
python-dotenv = "^1.0.1"
openai = "^1.54.4"
streamlit = "^1.40.1"
plotly = "^5.24.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
